move seg_profile into spatial_indexes
